+++ internal/ethapi/backend.go
@@ -90,7 +90,7 @@ type Backend interface {
 	ChainConfig() *params.ChainConfig
 	Engine() consensus.Engine
 	// CurrentTurnTerm return the turnTerm at the latest block
-	CurrentTurnTerm() (uint64, error)
+	CurrentTurnTerm() (uint8, error)
 
 	// This is copied from filters.Backend
 	// eth/filters needs to be initialized from this backend type, so methods needed by
